fix: Update devonfw-guide reference
Due to a probelmatic commit in devonfw-guide the auto update does not
work correctly. Therefore, a manual intervention is necessary.

In devonfw-guide the submodule  architecture-guidelines was not removed
completly. An update was done on devonfw.github.io, but the process
stopped, because of the wrong reference in devonfw-guide. Now all jobs
are failing and manual intervention is necessary.